refactor(workflow): move graph builder into core
